% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/districts.R
\docType{data}
\name{districts}
\alias{districts}
\title{districts data set}
\format{
A named integer vector with 13 elements:
\describe{
\item{Cheyenne Headquarters}{0}
\item{Pinedale Region}{1}
\item{Cody Region}{2}
\item{Sheridan Region}{3}
\item{Green River Region}{4}
\item{Laramie Region}{5}
\item{Lander Region}{6}
\item{Casper Region}{7}
\item{Nongame Program}{8}
\item{Migratory Game Bird Program}{9}
\item{Jackson Region}{10}
\item{Chapter 33}{33}
\item{Jackson Hole Nature Mapping}{35}
}
}
\usage{
districts
}
\description{
A named vector of district names and their corresponding numeric codes
used in the Wyoming Online Survey (WOS) data template.
}
\keyword{datasets}
