% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdgt.wos.R
\docType{package}
\name{spdgt.wos-package}
\alias{spdgt.wos}
\alias{spdgt.wos-package}
\title{spdgt.wos: Export Data for WOS}
\description{
This package provides functions to export data from SpeedGoat API and format it for import to the WOS database. This package is only useful to employees of Wyoming Game and Fish Department and is not intended for general consumption.
}
\author{
\strong{Maintainer}: Josh Nowak \email{josh.nowak@speedgoat.io}

Other contributors:
\itemize{
  \item SpeedGoat [copyright holder, funder]
}

}
\keyword{internal}
