% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos_export.R
\name{wos_export}
\alias{wos_export}
\title{Read and format Wildlife Observation System (WOS) data}
\usage{
wos_export(species, survey_type, analysis_unit, bio_year, observer, district)
}
\arguments{
\item{species}{the name of the species targeted by the survey}

\item{survey_type}{the type of survey conducted; must match one of the
survey types in the \code{survey_types} data set for the species}

\item{analysis_unit}{the name of the analysis unit surveyed}

\item{bio_year}{the biological year of the survey, for the winter of
2024-2025 the year of the survey is 2024}

\item{observer}{the name of the observer in "First Last" format}

\item{district}{the district name as found in the \code{districts} data set}
}
\value{
A data frame formatted for WOS import
}
\description{
Read and format Wildlife Observation System (WOS) data
}
\details{
This function combines \code{wos_read()} and \code{wos_format()} to read observations
from the database and format them for WOS import in a single step.

It performs input validation on all parameters to ensure they are correctly
formatted and valid. If no observations are found for the specified
parameters, a warning is issued and an empty WOS template data frame is
returned.

For more details check the documentation for \code{wos_read()} and \code{wos_format()}.
}
\examples{
\dontrun{
# Read and format some mule deer composition data for WOS
md_wos <- wos_export(
  species = "Mule Deer",
  survey_type = "Composition",
  analysis_unit = "Paintrock 207",
  bio_year = 2021,
  observer = "Jane Doe",
  district = "Pinedale Region"
)
}
}
