% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos_format.R
\name{wos_format}
\alias{wos_format}
\title{Format observations for entry in WOS}
\usage{
wos_format(x, observer, district, survey_type)
}
\arguments{
\item{x}{data frame of observations as returned by \code{wos_read()}}

\item{observer}{the name of the observer in "First Last" format}

\item{district}{the district name as found in the \code{districts} data set}

\item{survey_type}{the type of survey conducted; must match one of the
survey types in the \code{survey_types} data set for the species in \code{x}}
}
\value{
A data frame formatted for WOS import
}
\description{
Format observations for entry in WOS
}
